+++ core/che-core-api-core/src/main/java/org/eclipse/che/api/core/ErrorCodes.java
@@ -9,7 +9,7 @@
  * Contributors:
  *   Red Hat, Inc. - initial API and implementation
  */
-package org.eclipse.che.multiuser.resource.api.workspace;
+package org.eclipse.che.api.core;
 
 /**
  * Defines error codes, defined codes MUST NOT conflict, error codes must be in range <b>10000-14999</b> inclusive.
